test: split tests into unit and integration, add shared helpers
