+++ tests/unit/test_decomp_mpi.cpp
@@ -3,7 +3,7 @@
 
 #include "decomp.hpp"
 
-TEST(DecompMPI, GridDimsAndNeighbors) {
+TEST(Unit_Decomp, GridDimsAndNeighbors) {
     int world_size = 0, world_rank = -1;
     MPI_Comm_size(MPI_COMM_WORLD, &world_size);
     MPI_Comm_rank(MPI_COMM_WORLD, &world_rank);
