+++ tests/unit/test_field.cpp
@@ -2,12 +2,12 @@
 
 #include "field.hpp"
 
-TEST(Field, AllocationAndSize) {
+TEST(Unit_Field, AllocationAndSize) {
     Field f(4, 3, 1, 1.0, 1.0);
     ASSERT_EQ(f.data.size(), static_cast<size_t>(f.nx_total() * f.ny_total()));
 }
 
-TEST(Field, IndexingLayout) {
+TEST(Unit_Field, IndexingLayout) {
     Field f(2, 2, 1, 1.0, 1.0);
     for (int j = 0; j < f.ny_total(); ++j)
         for (int i = 0; i < f.nx_total(); ++i) f.at(i, j) = 10 * j + i;
@@ -18,7 +18,7 @@ TEST(Field, IndexingLayout) {
     EXPECT_DOUBLE_EQ(f.at(3, 3), 33);
 }
 
-TEST(Field, OutOfBoundsThrows) {
+TEST(Unit_Field, OutOfBoundsThrows) {
     Field f(4, 4, 1, 1.0, 1.0);
     EXPECT_THROW((void)f.at(-1, 0), std::out_of_range);
     EXPECT_THROW((void)f.at(f.nx_total(), 0), std::out_of_range);
