+++ tests/unit/test_io.cpp
@@ -17,7 +17,7 @@ static std::string cfg_path(const char* fname) {
 #endif
 }
 
-TEST(IO_Yaml, LoadsNestedBlocksAndBC) {
+TEST(Unit_IO_Yaml, LoadsNestedBlocksAndBC) {
     SimConfig cfg = load_yaml_file(cfg_path("dev.yaml"));
     EXPECT_GT(cfg.nx, 0);
     EXPECT_GT(cfg.ny, 0);
@@ -26,7 +26,7 @@ TEST(IO_Yaml, LoadsNestedBlocksAndBC) {
     EXPECT_GE(cfg.D, 0.0);
 }
 
-TEST(IO_CLI, SimpleScalarOverrides) {
+TEST(Unit_IO_CLI, SimpleScalarOverrides) {
     // base config comes from YAML
     auto tmpfile = "tmp_test.yaml";
     {
@@ -50,7 +50,7 @@ TEST(IO_CLI, SimpleScalarOverrides) {
     EXPECT_EQ(merged.output_prefix, "from_cli");     // CLI wins
 }
 
-TEST(IO_CLI, ICOverridesTakePrecedence) {
+TEST(Unit_IO_CLI, ICOverridesTakePrecedence) {
     std::vector<std::string> args = {"--ic.mode=preset",
                                      "--ic.preset=constant_zero",
                                      "--ic.A=999.0",
@@ -67,7 +67,7 @@ TEST(IO_CLI, ICOverridesTakePrecedence) {
     EXPECT_DOUBLE_EQ(merged.ic.yc_frac, 0.2);
 }
 
-TEST(IO_BC, ParseRoundtrip) {
+TEST(Unit_IO_BC, ParseRoundtrip) {
     EXPECT_EQ(bc_from_string("dirichlet"), BCType::Dirichlet);
     EXPECT_EQ(bc_from_string("neumann"), BCType::Neumann);
     EXPECT_EQ(bc_from_string("periodic"), BCType::Periodic);
@@ -77,22 +77,22 @@ TEST(IO_BC, ParseRoundtrip) {
     EXPECT_EQ(bc_to_string(BCType::Periodic), std::string("periodic"));
 }
 
-TEST(IO_CLI, InvalidBoundaryConditionThrows) {
+TEST(Unit_IO_CLI, InvalidBoundaryConditionThrows) {
     std::vector<std::string> args = {"--bc=foobar"};
     EXPECT_THROW({ merged_config(std::nullopt, args); }, std::runtime_error);
 }
 
-TEST(IO_CLI, InvalidGridSizeThrows) {
+TEST(Unit_IO_CLI, InvalidGridSizeThrows) {
     std::vector<std::string> args = {"--nx=-10", "--ny=128"};
     EXPECT_THROW({ merged_config(std::nullopt, args); }, std::runtime_error);
 }
 
-TEST(IO_CLI, InvalidTimestepThrows) {
+TEST(Unit_IO_CLI, InvalidTimestepThrows) {
     std::vector<std::string> args = {"--dt=0.0", "--steps=10"};
     EXPECT_THROW({ merged_config(std::nullopt, args); }, std::runtime_error);
 }
 
-TEST(IO_CLI, InvalidICPresetThrows) {
+TEST(Unit_IO_CLI, InvalidICPresetThrows) {
     SimConfig cfg;
     cfg.ic.mode = "preset";
     cfg.ic.preset = "notarealpreset";
